AcademicRAGProject/requirements.txt
streamlit
PyMuPDF
matplotlib
networkx
sentence-transformers
transformers
scikit-learn
faiss-cpu
